+++ packages/frontend/src/pages/Settings.tsx
@@ -16,6 +16,7 @@ import {
     IconUserCircle,
     IconUserPlus,
     IconUsers,
+    IconUserShield,
 } from '@tabler/icons-react';
 import { FC } from 'react';
 import { Redirect, Route, Switch } from 'react-router-dom';
@@ -39,6 +40,7 @@ import ProfilePanel from '../components/UserSettings/ProfilePanel';
 import ProjectManagementPanel from '../components/UserSettings/ProjectManagementPanel';
 import SlackSettingsPanel from '../components/UserSettings/SlackSettingsPanel';
 import SocialLoginsPanel from '../components/UserSettings/SocialLoginsPanel';
+import UserAttributesPanel from '../components/UserSettings/UserAttributesPanel';
 import UserManagementPanel from '../components/UserSettings/UserManagementPanel';
 import { useOrganization } from '../hooks/organization/useOrganization';
 import { useActiveProjectUuid } from '../hooks/useActiveProject';
@@ -188,7 +190,25 @@ const Settings: FC = () => {
                                         }
                                     />
                                 )}
-
+                                {localStorage.getItem('USER_ATTRIBUTES') &&
+                                    user.ability.can(
+                                        'manage',
+                                        subject('Organization', {
+                                            organizationUuid:
+                                                organization.organizationUuid,
+                                        }),
+                                    ) && (
+                                        <RouterNavLink
+                                            label="User attributes"
+                                            to="/generalSettings/userAttributes"
+                                            exact
+                                            icon={
+                                                <MantineIcon
+                                                    icon={IconUserShield}
+                                                />
+                                            }
+                                        />
+                                    )}
                                 <RouterNavLink
                                     label="Appearance"
                                     exact
@@ -410,6 +430,17 @@ const Settings: FC = () => {
                 {user.ability.can('view', 'OrganizationMemberProfile') && (
                     <Route path="/generalSettings/userManagement">
                         <UserManagementPanel />
+                    </Route>
+                )}
+
+                {user.ability.can(
+                    'manage',
+                    subject('Organization', {
+                        organizationUuid: organization.organizationUuid,
+                    }),
+                ) && (
+                    <Route path="/generalSettings/userAttributes">
+                        <UserAttributesPanel />
                     </Route>
                 )}